+++ libs/optixtracer/c_src/mogTracingBwd/params.h
@@ -37,7 +37,7 @@ struct MoGTracingBwdParams
     float slabSpacing;
 
     float hitMinGaussianResponse;
-    float sphDegree;
+    unsigned int sphDegree;
     uint2 frameBounds;
     
     unsigned int frameNumber;
